refactor: RedisUtil 패키지 이동
