+++ src/test/java/com/petit/toon/repository/util/RedisUtilTest.java
@@ -1,5 +1,6 @@
 package com.petit.toon.repository.util;
 
+import com.petit.toon.util.RedisUtil;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
